Fix the Unsubscribe timeout problem
The wait Unsubscribe now is sync so it is removed from the consumers list correctly.
The connection can be closed correctly since there are not pendig consumers.

Fix the handle deliver in case of the consumer is removed but there are still chunks on the wire.
Add Degug Asserts to validate the buffer size, it can help to understand some parse chunk error.
That can be temporany at some point we can remove them.
Ref https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/issues/310

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>