Add Info class to the Producer/Consumer classes
The call returns the Reference and Stream.
Note that for the producer class the Reference is deprecated and it will be removed
at some point. So the info.reference for the producer will be removed as well.

We can add other info about the consumer and the consumer in the future.

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>