Improve the super stream reconnection (#344)
Fixes: https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/discussions/333
Fixes: https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/issues/335

# Preface
The commit is focused on making the client more stable during the server restart and better handling the metadata update event.

Most of the changes are in the super stream producer and consumer.
See the issue https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/discussions/333

# 1- What's changed in the super stream part


The RawSuperStreamProducer and RawSuperStreamConsumer expose two callbacks:
- `ConnectionClosedHandler (reason, stream)` 
- `MetadataHandler(update)` 
- `ISuperStreamProducer` and `ISuperStreamConsumer` interfaces 

ConnectionClosedHandler
--
The callback is raised each time a partition is closed for some reason. You can check the reason with:
```
1- public const string Normal = "TCP connection closed normal";
2- public const string Unexpected = "TCP connection closed unexpected";
```

1- The super stream partition is closed by the `Close()` method, so there are no problems. It can be helpful in terms of alerting or logs.

2- The super stream partition is closed in an unexpected way like kill the connection, network problems broker restart etc..

In that case, you can reconnect the partition with the `ReconnectPartition` method.

MetadataHandler
---
It happens when there is a change in the stream topology.

In this case, the server drops all the producers and consumers. So, you need to react to the Metadata Handler event.  You can decide to reconnect the producer and consumer
using the `ReconnectPartition` 

Producer and Consumer classes
--

The `Producer` and `Consumer` classes handle automatically the `ConnectionClosedHandler` and the `MetadataHandler` events for stream and super stream.


# 2- Fail fast

The Producer and Consumer classes can restore the connection automatically when they are up and running.
The classes will fail immediately if there are problems during the first connection.
This behaviour is now like the Java Client's.


---------

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>