Add MissingHeartbeat to the possible Connection Closed Reasons
to the Connection class.
Remove the update status form the the Dispose() function.
It is not necessary since the dispose can be called only from the Client close() function
where the status is set.

Add a function to convert the connetion close reason string to the Reliable.ChangeStatusReason Enum.
This mapping should be improved to the version 2.0 where we can introduce some breaking changes.
In the version 1.x we need the conversion function.

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>