Use env to determine bash location
This is important in Darwin (MacOS) because `/bin/bash` is the system
Bash, which is version 3.x. This version of bash does not recognise the
test operator `-v`, therefore the script breaks in line `if [[ -v foo]]`

With this change, `env` picks up the Bash version installed from
Homebrew (if available), which is Bash 5.x; this version has support for
`-v` operator.